POWER BI PROJECT BY GETRUDE ODINAKACHI
This is a project that analyses the growth and challenges of a fashion brand.   I created a dashboard that will help showcase my findings in the dataset.   Generally, this project was an excellent opportunity to apply what I learnt in the training and gain hands-on-experience with a real world data.